Figured out how to get the oauth signature to pass from the client to the wp backend proerly.
This is the beginning of posting to the API, it creates the book successfully now. The one problem is that I had to expose the client secret to the front end. I decided to do this after authorization is complete, to at least decrease the possibility of an attacker getting it, however this is NOT a production ready solution.

Would likely need to run all requests through a local passthru to do this correctly.