Major revamp of the backend login system.
In order to avoid having to continue creating auth tokens, I've decided to break the rules and implement a local user storage for accounts. This isn't ideal, since the rules were to not have any local database and while this counts, the only other solution was to create temporary oauth1 access tokens that die with the user's session. This way allows for longer term sessions to continue to exist, without the need for a full database.

This solution also includes a small registration feature to create an account.