Make the likes a logged in only feature.
This is not idea, since it basically means you're liking your own post, and it means likes are sort of useless. Ideally, we would be able to write and record likes and data from anyone. However, the WP-API thrives best when dealing with WP users and setting up global write stores, are not really the WP-API's ideal use case.